@@ -28,9 +28,9 @@ func (r *InMemoryRepository) DeleteURLS(ctx context.Context, userID string, shor
 }
 
 func (r *InMemoryRepository) SaveBatchURLS(userID string, links types.BatchLinks) (types.ResponseBatch, error) {
-	var response types.ResponseBatch
-	for _, v := range links {
-		response = append(response, types.ResponseBatchJSON{CorrelationID: v.CorrelationID, ShortURL: v.ShortURL})
+	response := make(types.ResponseBatch, len(links)) // allocate required capacity for the links
+	for i, v := range links {
+		response[i] = types.ResponseBatchJSON{CorrelationID: v.CorrelationID, ShortURL: v.ShortURL}
 	}
 	return response, nil
 }
@@ -44,17 +44,18 @@ func (r *InMemoryRepository) GetURL(shortURL string) (types.OriginalLink, error)
 }
 
 func (r *InMemoryRepository) GetUserStorage(userID string) ([]types.Link, error) {
-	var links []types.Link
 	ids, ok := r.inMemoryUserStorage[userID]
 	if !ok {
-		return links, errors.New("UserID not found")
+		return nil, errors.New("UserID not found")
 	}
-	for _, v := range ids {
+
+	links := make([]types.Link, len(ids)) // allocate required capacity for the links
+	for i, v := range ids {
 		URL, ok := r.inMemoryMap[v]
 		if !ok {
 			return links, errors.New("ID not found")
 		}
-		links = append(links, types.Link{ShortURL: v, OriginalURL: URL})
+		links[i] = types.Link{ShortURL: v, OriginalURL: URL}
 	}
 	return links, nil
 }
@@ -63,6 +64,7 @@ func (r *InMemoryRepository) Ping() bool {
 	return true
 }
 
+// NewInMemoryRepository returns a new InMemoryRepository.
 func NewInMemoryRepository() *InMemoryRepository {
 	log.Print("Memory storage is used")
 	return &InMemoryRepository{inMemoryMap: make(map[string]string), inMemoryUserStorage: make(map[string][]string)}
